+++ src/aslm/view/main_window_content/display_notebook.py
